classes/image_populate_mfgtool: init
The class provides the infrastructure for MFGTOOL generation and is tied
to images. To generate the bundle, the task populate_mfgtool must be
called. For example:

,----[ Running populate_mfgtool for core-image-minimal image ]
| $: bitbake core-image-minimal -c populate_mfgtool
`----

The class behavior is controlled through the MFGTOOLSCRIPT variable. The
MFGTOOLSCRIPT variable itself specifies a space-separated list of the
scripts to enable. Following the scripts, you can determine the behavior
of each script by providing up to two order-dependent arguments, which
are separated by commas. You can omit any argument you like but must
retain the separating commas. The order is important and specifies the
following:

  1. Extra dependencies that should be added to the do_populate_mfgtool
     task, if the script is enabled.
  2. Extra binaries that should be added to the bundle, if the script is
     enabled.

For example:

,----[ Defining bootloader.uuu.in script ]
| MFGTOOLSCRIPT = "bootloader.uuu.in"
| MFGTOOLSCRIPT[bootloader.uuu.in] = "dep1:do_deploy dep2:do_deploy,file1 file2"
`----

The virtual/bootloader:do_deploy is added to do_populate_mfgtool
dependencies and ${UBOOT_BINARY} copied to the bundle, only if the
script is enabled.

During the mfgtool bundle generation, the uuu.in files are processed and
some variables replaced. The variables are:

  - MACHINE
  - UBOOT_BINARY
  - SPL_BINARY
  - IMAGE_BASENAME

Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>